+++ strucmotif-search-core/src/main/java/org/rcsb/strucmotif/io/AssemblyInformationProvider.java
@@ -10,4 +10,10 @@ public interface AssemblyInformationProvider {
      * @return a map with operator expressions as key and a collection of all corresponding assembly ids as value
      */
     Map<String, Set<String>> selectAssemblyMap(String structureIdentifier);
+
+    /**
+     * If dealing with undefined assemblies: What's their identifier?
+     * @return access to the `undefined-assembly-identifier` config property
+     */
+    String getUndefinedAssemblyIdentifier();
 }